Add a SECURITY.md file
... containing information about the security bounty program and how to report security issues.

The file is placed in the `.github` repository, which means that it will automatically be used by all repos which don't have an explicit `SECURITY.md` file in the repo itself.

Once this PR has been merged, the `SECURITY.md` file in those repos which currently have one can be removed, so all repos display the same security information.

Refs:
* https://docs.github.com/en/code-security/getting-started/adding-a-security-policy-to-your-repository
* https://docs.github.com/en/communities/setting-up-your-project-for-healthy-contributions/creating-a-default-community-health-file