QA: add reusable merge conflict check workflow
This commit introduces a new, reusable workflow to check if PRs are in a "conflict state", and if so, add a "merge conflict" label and leave a comment on the PR.

The workflow uses a pre-existing action runner which offers a number of options to configure the action.

While the workflow itself is pretty straight-forward, it has been set up as a reusable workflow to centrally manage the default option settings for all repos.

Individual calling workflows can still overrule these defaults for the most pertinent options.

For the less pertinent settings, this overload has not been enabled, but these options can still be opened up to calling workflows at a later point in time if there is a need for it.

Refs:
* https://github.com/eps1lon/actions-label-merge-conflict